feat(회원) : policy table을 사용한 테스트코드 수정  (#139)

+++ src/main/java/com/project200/undabang/member/entity/Member.java
@@ -85,7 +85,7 @@ public class Member {
                 .build();
     }
 
-    public void decreaseMemberScore(Byte decreaseScore){
+    public void decreaseMemberScore(int decreaseScore){
         if(this.memberScore > 0){
             this.memberScore = (byte) Math.max(0, this.memberScore - decreaseScore);
         }
